requirements.txt
greenlet==2.0.2
loguru==0.6.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pydantic==1.10.5
python-dotenv==1.0.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
